backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==1.0.3
python-multipart==0.0.6
websockets==12.0